add xor parser and put parser in individual files
